Consolidate into seekbar patch and settings menu
