+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/LivestreamDvrPatch.kt
@@ -5,7 +5,7 @@
  * See the included NOTICE file for GPLv3 §7(b) and §7(c) terms that apply to this code.
  */
 
-package app.morphe.patches.youtube.video.livedvr
+package app.morphe.patches.youtube.interaction.seekbar
 
 import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
@@ -13,17 +13,16 @@ import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
 import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
 import app.morphe.patches.youtube.misc.settings.PreferenceScreen
 import app.morphe.patches.youtube.misc.settings.settingsPatch
-import app.morphe.patches.youtube.shared.Constants.COMPATIBILITY_YOUTUBE
+import app.morphe.patches.youtube.video.livedvr.VideoStreamingDataAllowSeekingFingerprint
 import app.morphe.util.addInstructionsAtControlFlowLabel
 import app.morphe.util.findInstructionIndicesReversedOrThrow
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 
-private const val EXTENSION_CLASS_DESCRIPTOR = "Lapp/morphe/extension/youtube/patches/LiveDvrPatch;"
+private const val EXTENSION_CLASS_DESCRIPTOR = "Lapp/morphe/extension/youtube/patches/LivestreamDvrPatch;"
 
 @Suppress("unused")
-val liveDvrPatch = bytecodePatch(
-    name = "Live DVR",
+val livestreamDvrPatch = bytecodePatch(
     description = "Enables video seeking on live streams that have disabled DVR.",
 ) {
     dependsOn(
@@ -31,11 +30,9 @@ val liveDvrPatch = bytecodePatch(
         settingsPatch,
     )
 
-    compatibleWith(COMPATIBILITY_YOUTUBE)
-
     execute {
-        PreferenceScreen.VIDEO.addPreferences(
-            SwitchPreference("morphe_force_live_dvr")
+        PreferenceScreen.SEEKBAR.addPreferences(
+            SwitchPreference("morphe_livestream_dvr")
         )
 
         VideoStreamingDataAllowSeekingFingerprint.method.apply {
@@ -45,7 +42,7 @@ val liveDvrPatch = bytecodePatch(
                 addInstructionsAtControlFlowLabel(
                     returnIndex,
                     """
-                        invoke-static { v$returnRegister }, $EXTENSION_CLASS_DESCRIPTOR->enableLiveDvr(Z)Z
+                        invoke-static { v$returnRegister }, $EXTENSION_CLASS_DESCRIPTOR->enableLivestreamDvr(Z)Z
                         move-result v$returnRegister
                     """
                 )
