+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/SeekbarPatch.kt
@@ -8,13 +8,16 @@ val seekbarPatch = bytecodePatch(
     name = "Seekbar",
     description = "Adds options to disable precise seeking when swiping up on the seekbar, " +
             "slide to seek instead of playing at 2x speed when pressing and holding, " +
-            "tapping the player seekbar to seek, and hiding the video player seekbar."
+            "tapping the player seekbar to seek, " +
+            "hiding the video player seekbar, " +
+            "and enabling seeking in livestreams."
 ) {
     dependsOn(
         disablePreciseSeekingGesturePatch,
         enableSlideToSeekPatch,
         enableTapToSeekPatch,
         hideSeekbarPatch,
+        livestreamDvrPatch
     )
 
     compatibleWith(COMPATIBILITY_YOUTUBE)
