+++ patches/src/main/resources/addresources/values/youtube/strings.xml
@@ -543,6 +543,11 @@ Verify the package name is correct and the app is installed"</string>
     <string name="morphe_disable_precise_seeking_gesture_summary_on">Precise seeking gesture is disabled</string>
     <string name="morphe_disable_precise_seeking_gesture_summary_off">Precise seeking gesture is enabled</string>
 
+    <!-- interaction.seekbar.livestreamDvrPatch -->
+    <string name="morphe_livestream_dvr_title">Enable livestream DVR</string>
+    <string name="morphe_livestream_dvr_summary_on">Seeking is allowed on all live streams</string>
+    <string name="morphe_livestream_dvr_summary_off">Seeking may not be available on live streams</string>
+
     <!-- interaction.seekbar.enableSlideToSeekPatch -->
     <string name="morphe_slide_to_seek_title">Enable slide to seek</string>
     <string name="morphe_slide_to_seek_summary_on">Slide to seek is enabled</string>
@@ -1723,11 +1728,6 @@ Limitations:
 • Video playback will use more internet data than VP9 or AV1
 • HDR videos will not use AVC
 • Some devices cannot force AVC"</string>
-
-    <!-- video.livedvr.liveDvrPatch -->
-    <string name="morphe_force_live_dvr_title">Enable live DVR</string>
-    <string name="morphe_force_live_dvr_summary_on">Seeking is allowed on all live streams</string>
-    <string name="morphe_force_live_dvr_summary_off">Seeking may not be available on live streams</string>
 
     <!-- video.quality.advancedVideoQualityMenuPatch -->
     <string name="morphe_advanced_video_quality_menu_title">Show advanced video quality menu</string>
